Change how "gazctl {journals,shards} apply" takes stdin
Previously, `journals apply` and `shards apply` subcommands read from
stdin if the `--specs` option was not provided. This lead to the command
appearing hung if a user passed a file name but forgot to use `--specs`.
Stdin would be used and the filename, given as an arg, would be silently
ignored.

Taking a cue from `kubectl apply -f`, the `--specs` option is now
required. `--specs -` will read from stdin.