Relax etcd revision check on KeySpace#Load (#188)
The broker's initial load of the KeySpace after announcing itself to the
cluster was racing the allocator's attempt to schedule it.

KeySpace#Load uses a mirror.Syncer to receive a range of keys over a
channel of type clientv3.GetResponse. It was assumed that the response
revision would be identical to the revision in the request. This turns
out to be false; it represents the revision of the key-value store when
generating the [response][1]. Therefore, it's possible for the revisions
in a sync to increase over a single load attempt.

In fact, this is common in practice because a broker announces itself
prior to calling KeySpace#Load triggering the allocator to schedule it,
writing the new schedule to etcd and changing the revision of the
key-value store.

cabc807 was the first attempt to address this, but was flawed because it
required each response of the sync to be strictly greater than the
previous.

[1]: https://coreos.com/etcd/docs/latest/learning/api.html#response-header