Remove extraneous warnings from GCS backed stores
With this change, gcsBackend#List ignores objects that represent
directories when iterating through the objects of a bucket for a
journal. Such objects never represent fragment files and attempting to
parse their names as such will almost certainly fail, writing a warning
to the logs when, in fact, everything is as expected.