broker: roll spool after very first journal write
Spools having Begin == 0 and End != 0 are now proposed to be immediately
rolled, regardless of target journal fragment length. This protects
against data loss by quickly "dirtying" the fragment store, even for
journals which are written slowly and would ordiniarly not persist a
spool anytime soon.

If Etcd consistency is lost, a dirty fragment store informs recovering
brokers that the journal has had previous writes, and that an explicit
`gazctl reset-head` should be performed to avoid data-loss. If the store
were left prestine, it's impossible to distinguish the "recovery" case
from the "brand new journal" case, and it's likley data-loss will occur
as journal offsets are double-written.

Also, fix a missing lock in Persister.