keyspace: use WithRequireLeader in Watch, and retry
With this setting, if our watched Etcd instance is partitioned from its
majority it will abort our watch stream rather than wait for the
partition to resolve (the default). We'll then retry against another
endpoint in the majority.

Without this setting, there's a potential deadlock case if our
watch Etcd is partitioned from the majority, but *we* are not and
can keep our lease alive. Specifically this can lead to our being
allocator leader but being unable to observe updates.