server: add GRPCLoopback
Test and production servers both want to build loopback connections to
themselves, for use with protocol.dispatcher. Consolidate the proactive
building of these connections into server.Server.

Automatic graceful stopping of the server is removed from server, and is
now an explicit responsibility of broker.Service / consumer.Service.
This provides better clarity on when and why the gRPC server is stopped,
and requires that tests be explicit in tearing down resources they
allocate (a good thing).