protocol: tweak dispatcher SubConn handling
Don't sweep the default service SubConn, or return TransientFailure if
it's not presently available (instead, we should block till it comes
back).

Don't allow a broken SubConn to come back into service simply because
it enters the Connecting state again. It must come back fully or not at
all, and until then continue to return TransientFailure from Pick().

If we don't require a specific process in another zone, prefer to wait
for a same-zone member to be ready rather than dispatch across zones.
Eg, if a journal assignment is changing, we'd rather wait for a member
in our zone (which should be fast) than start a very long lived read
across zones (which costs more).

Also update etcd import, define a PageLimit maximum value, and factor
out JournalRouteMatchesAssignment for future use in `broker` package.