consumer: better shutdown behavior and FAILED shard
Resolver has an explicit stopServingLocalReplicas() which is invoked if
the task.Group context is cancelled. This allows the server to begin
cancelling any related replica RPCs so that GracefulStop can succeed.

Service now manages server GracefulStop (paralleling how broker.Service
handles it).

FAILED shards are now consistent if the primary has also failed. This
allows the failed replica to swap in for the primary, such that the
primary can exit or be re-assigned.

Update etcd imports.