store-rocksdb: new package for RocksDB recording and other details
The package implements a "hooked" rocksdb::Env for observing and
recording the file operations performed by an embeded RocksDB instance.

It also offers an ArenaIterator for iteration over a RocksDB that
ammortizes CGO calls using a memory arena for pre-fetching.

This functionality is hoisted out and updated from a gorocksdb branch I
had been maintaining. By placing these bits here, we should be able to
revert to mainline gorocksdb.

This is also a first step of hoisting RocksDB out of the `consumer` package
and dropping the `rocksdb` build tag.