broker/client: changes for exactly-once semantics
AppendService is refactored to take an entire AppendRequest (rather than
just a journal). This allows for passing through *CheckRegisters.

OpFuture interface is defined, and extracted from AsyncAppend (which
implements it). This will allow for the `consumer` package to consider
other types of futures (like external RDBMS commits) in addition to
AsyncAppends. Likewise, StartAppend can now depend on an OpFuture of any
type (not necessarily an AsyncAppend). AppendService tracks errors
across serveAppends(), and _any_ dependency error of a journal will
permanently fail all further appends to that journal.

serveAppends is tuned to zero-out many AsyncAppend fields which are no
longer needed once an AsyncAppend completes. This lets an *AsyncAppend
be retained for a long period of time without it pinning a whole bunch
of other objects from the garbage collector.

Also:

* GetJournal is added.
* RetryReader exports its ctor argument fields to allow reuse of its
  client & context.

Issue #210