recoverylog: refactor RecordCreate into RecordOpenFile
RecordOpenFile takes flags having the same meaning as os.OpenFile, and
RecordCreate now calls it.

Rather than returning a FileRecorder, RecordOpenFile directly returns
the recorded FNode. FileRecorder is still available as an explicit
convienence for wrapping an FNode.

Minor refactors of the AFero and RocksDB records to accomodate this
change.

Issue #212