consumer,gazctl: verify referenced journals of applied ShardSpecs
Upsert'd ShardSpecs are verified to ensure that referenced journals
exist, and have appropriate content-types.

This check is done both by the client (as a pre-flight) and the server
(with each applied batch).

This mitigates a current foot-gun whereby a user can create ShardSpecs
which will immediately fail, because journals haven't been created or
are miss-spelled. It also ensures K8s test jobs which apply specs also
run in the right order.

Issue #216