message: introduce Framing registry and add CSV support
Framing instances are now dynamically registered and loaded at
run-time. This allows applications to register new types of bespoke
Framings (eg, for Thrift) which are not included in the core repository.

A new framing for the `text/csv` content-type is also added.

The Framing interface is re-worked to make it easier to implement
framed message readers which want to keep state in between read
frames.

The Unpack / Unmarshal distinction is also removed, as it's
probably not a good idea to have Gazette attempt to automatically
handle corrupt messages in the journal stream (should they be skipped?
does the binary need to be updated?). This was a helpful tool when
Gazette was less mature, but now it's better to get an operator
involved to explicitly diagnose and recover.