consumer: add ShardsWithSource, godocs, change SQLStore placeholders
ShardsWithSource indexes over ShardSpecs having a given journal as a
source. It addresses a common need of applications wanting to map from
an app-specific key, to a journal partition, to a responsible shard.

Fix a bug in verification of referenced journals of ShardSpecs (that
it's okay if a spec omits a recovery log).

SQLStore switches to using PostgreSQL-compatible $N placeholders
instead of '?'.