cmd: make fewer flags required, support --broker.file-root, & remove demo mode
file-root flag lets broker and consumer applications use a shared NAS
file root. Also tweak brokers and consumers to bind a random port by
default, to generate a random instance ID, and to generate a default
consumer Etcd prefix (basically, require as few flags to be specified
as possible, where it's possible to instead come up with reasonable
dynamic defaults).

Demo mode is removed as it's weird, and the broker tutorial can be
updated to include just a couple more steps which get to equivalency.