allocator: refactor ItemValue.IsConsistent into a free function callback
This allows `broker/protocol.JournalSpec.IsConsistent` to be hoisted out
of the `protocol` package and into the `broker` package instead, which
is a step towards removing all dependency on the `keyspace` package
(which in turn depends on `etcd`) from the `protocol` package.

A similar refactor is done for ShardIsConsistent which removes
dependence on `keyspace` & `etcd` from `consumer/protocol` as well.