broker: remove broken response-time metric, add started & completed
Current response-time metric is broken due to an integer rounding bug,
which I take as positive proof that nobody is currently using it.

Remove it, and rely in gRPC-prometheus histogram instrumentation instead.

Do add RPC starts & completions, broken out by operation and status.
This can also be used to snapshot the # of current live RPCs.