testing: testify/assert => testify/require
It's common in gazette tests that an earlier assertion failure can
cause a later test deadlock. The root cause of the failure is lost
unless tests are diligent about using require. In practice, seeing
multiple failures from a test is almost never actually valuable, so
switch to require across the board.