Change how rocksdb is build and linked
Rocksdb was previously built by our makefile, and downstream consumers
would often share this build logic by importing the `mk/common-build.mk`
file into their own Makefiles. This creates a bit of a mess, because any
changes to `mk/common-build.mk` would impact untold numbers of consumer
applications. This commit begins to change that paradigm by moving
rocksdb out of our makefile and into the ci-builder docker image. Now,
rocksdb is treated the same as any other native library dependency. It
is expected to be provided by the build environment. If a consumer
wishes to build with a specific rocksdb version, they can simply install
whatever version they like into their build environment. To make this
easy, the `ROCKSDB_VERSION` argument was added to `ci-builder.Dockerfile`.
Consumers can use that to easily get a build environment setup with
whatever version they like.

Additionally, an option was added to allow rocksdb and other native
libraries to be statically linked. This currently requires linker args
to be specified in the `CGO_LDFLAGS` environment variable, but this
requirement may be lifted with some additional work in the gorocksdb
project. As of this PR, static linking is the default for for all the
gazette examples, and the release docker image will no longer include
shared versions of rocksdb or any of the compression libraries.