broker: re-work journal '?' queries into a ';metadata/path/segment'
By making these metadata suffixes a URL-like path segment,
this change clarifies that Journals -- including any metadata
path suffix -- are to be encoded as URLs within the Path
component (only).

A notion of Journal queries is too confusing, because Journals are not
actually URLs (and are not trying to be URLs), and it raises a host of
questions around encoding and inter-mixing with actual URL query arguments.