recoverylog: require that the last segment of FSMHints match its Log
It must not be the case that FSMHints, when flattened, have a final
log segment which reads a different log than that of the hints.

The rationale is that playback cannot know the appropriate SeqNo and
Checksum to expect when reading from the FSMHints.Log. If, for example,
there's a no-op Operation sequenced _after_ the final segment of the
hints, the player cannot know whether that operation should or
shouldn't be applied before starting to read from the FSMHints.Log.

Effectively this constraint is asserting that there *must* be at least
one live file operation tracked in the FSMHints.Log, and it must be
the final segment of the FSMHints. It's presence ensures that playback
knows an appropriate SeqNo/Checksum/Offset to begin reading that log
from.