consumer: shards make progress on duplicate / roll-back messages
Instrument message.Sequencer to provide an indication of whether a given
envelope continues a pending sequence, or is an ack / duplicate /
roll-back.

Have txnRead use this bool to determine whether polling the sequencer is
required. Notably this avoids polling (and new progress signaling logic)
in the common case of many read-uncommitted messages which will begin a
transaction only once committed.

Then, for messages read outside of a transaction which are _not_ a
continuation (meaning they're a duplicate, ACK, or roll-back),
perform a signaled update of the read progress of the shard.

This aids a shard Stat RPC which may be ongoing, and which may have
provided read-through offsets which include the duplicate message.
Without this behavior, the read progress wouldn't be recorded and the
Stat RPC could hang indefinitely if the last message of read journals
is a duplicate.