client: AppendService owns its buffer sync.Pool
Rather than using a global singleton Pool.

In the future this may allow for better coupling of resource quotas
and back-pressure against specific AppendService instances,
correlated to different consumer.Shards.

The immediate motivation is fixing a test-only bug, where a
mockFile{} fixture is added to the global Pool and causes a
future test case to fail.

Issue #297