consumer: new transaction test cases modeling partial dequeue
TestTxnMaxDurWithinDequeueSequence simulates a transaction maximum
duration timeout in the middle of processing a transactional sequence.
Expectation is that the current transaction closes, and the next transaction
picks up where it left off.

TestAppDeferWithinDequeueSequence simulates an application which defers
processing of a message to a future transaction. Similar expectation:
the current transaction closes and the next picks up where it left off.

Issue #300