message: rework Sequencer for finer-grain state tracking
Before, Sequencer expected that runs of acknowledged messages would be
processed as a single atomic unit, and only concerned itself with
tracking state changes from *uncommitted* messages.

This is now inverted: when a run of acknowledged messages is ready for
processing, Sequencer is Step()'d through each message in turn,
and may further be asked to produce a Checkpoint() at any time.

Checkpoints within acknowledged runs are accomplished by iteratively
tightening the ProducerState bound to reflect the Clock of the *last*
stepped-past message and the offset of the stepped-to *next* one.

QueueUncommitted provides a finer-grain QueueOutcome result type which
gives detail into the exact queue decision made, and the QueueOutcome is
now the mechanism by which the caller discovers the need to start a
replay read (rather than receiving a ErrMustStartReplay, which is
removed).

Finally, Sequencer now takes responsiblity for maintaining read-through
offsets of journals. This is a more natural fit because read-through
offsets *must not* reflect a message ACK until its acknoweldged run has
been fully processed, and Sequencer best understands this detail.

Issue #300