consumer: invert transaction loop to allow finer-grain boundaries
Formerly, processing acknowledged messages was done in a tight loop
within txnConsume. This was pretty problematic for large transactions,
because the application was required to handle with potentially far more
messages in a single transaction that it might prefer (or even have
resources to handle). It can also cause tranaction timers to be wildly
ignored, depending on how much work the application does per-message.

Now, processing of a run of acknowledged messages is hoisted to txnStep,
and includes timer checks. An unfinshed run of messages is able to be
deferred to a following transaction, and a new error is introduced to
also allow applications to make a determination of when a transaction
should close.

The updated message.Sequencer is used for tracking and production of
fine-grain Checkpoints, including ones which are within an acknowledged
message run.

Issue #300