add flags for globally bounding journal / shard replication & stores
It's very common to run a single broker or consumer instance in testing
contexts. However, a single process can never replicate a journal or
shard more than one time.

To date the practice has been to tweak journal / shard specifications so
that they "fit" inside a testing broker: limiting replication /
hot_standbys to one or zero (respectively), and not providing a remote
fragment store (so that data stays local and ephemeral).

The downside is this means that journal / shard specifications used
within a testing context must differ from those used in production.

Instead, introduce command-line flags which allow setting an effective
limit on replication (both journal and shard), and which disable the use
of any configured journal stores. This allows unmodified production specs
to be usefully applied and run within a single-process "cluster".