server: enable binding of unix domain sockets
Binding to UDSs is a better practice for hermetic, contained test
clustesr which don't pollute the port space of the host.

A Unix socket is specified as a `unix://path` or `unix://host/path`
port flag.

The address always maps into a protocol.ProcessSpec with Address of
`unix://host/path` (the host/authority is explicit and required).

As gRPC doesn't want an authority, add a new routine to map back into a
form that gRPC is willing to dial.