Avoid setting replica status in all tests
This test helper change caused several failures in tests I did not
touch. I believe this is caused by real behavior being triggered when
the shard has a "PRIMARY" replica, but the tests do not expect those
effects and thus fail.

So rather than setting the replica status for every test using the
`allocateShard` helper, we'll just be very explicit and set the status
in the places where it needs to be set.