Add `shard unassign` cmd (#312)
* Add Unassign RPC to consumer protocol

* Implement ShardUnassign RPC

* Add `shards unassign` command to gazctl

* Inline config definition

* Add more details to command output

* Check shard consistency before removing assignment

* Remove Header field from Unassign RPC

* Add simple RPC validation tests

* Add `--failed` flag to `shards unassign`

* Batch remove shard assignments

Submit multiple shard ids in a single RPC, rather than submitting
multiple RPCs.

* Gracefully skip non-failed replicas

* Avoid setting replica status in all tests

This test helper change caused several failures in tests I did not
touch. I believe this is caused by real behavior being triggered when
the shard has a "PRIMARY" replica, but the tests do not expect those
effects and thus fail.

So rather than setting the replica status for every test using the
`allocateShard` helper, we'll just be very explicit and set the status
in the places where it needs to be set.

* Output previous shard status after unassign cmd

The user can always run `shards list` immediately to see the new status.