Report which shards had assignments removed
Due to the nature of the `--failed` flag, we can't know which replicas
will be targeted by this command by the List selector alone. We need to
check the ReplicaStatuses. We're already doing that in the consumer
process, so just record and send back which shards actually had
assignments removed.