consumer: shard transitions account for delete-then-create race
If a shard assignment is removed (for example, because its FAILED)
and is then immediately re-created by the allocation coordinator,
we previously could fail to notice that the prior local shard was
invalidated and should be re-created.

Account for this by additionally monitoring the CreateRevision of
assignments and treating an assignment of the same or higher slot,
with a differing CreateRevision, as being a novel assignment which
requires a novel local shard to be boot-strapped.

Fixes #314