Bail if no shards will be unassigned
This fixes a bug which caused the shard supervisor to stall
indefinitely. If it ran `shards unassign` and there was nothing to be
unassigned, it would never increment the etcd revision with its
transaction. Then when we waited to observe our changes as reflected by
the etcd mod revision, we would end up waiting until the next arbitrary
transaction to complete. Needless to say, this is not the desired
behavior.