error when s3 client is missing region config
When the `region` configuration is missing the AWS S3 client
will fail essentially all operations. This is true even if the
`endpoint` is provided in the storage URL. Because the client is cached,
this same failure will continue to happen, even _after_ you update
`~/.aws/config` to set the `region`.

This commit adds an explicit check for the `Region` when initializing a
new S3 client, to prevent the invalid client configuration from being
persisted in the `clients` map.