Fix flakey store_sql_test
gazette/core#329 introduced a change to synchronously fail shards if the result of
`store.StartCommit` is immediately available. This is _technically_ a
race condition, since some asynchronous operations may still completed
and observed by `txnStartCommit`. This changes the test to only match
against a portion of the final error message, so that it tolerates the
error being returned either synchronously or asyncronously.