Fix etcd bootstrap in examples
The etcd server in the examples had a lot of troubles starting, and
would frequently fail to start, printing error output like:

```
Waiting for etcd-0.etcd to come up
Waiting for etcd-1.etcd to come up
Waiting for etcd-2.etcd to come up
Re-joining etcd member
cat: /var/run/etcd/member_id: No such file or directory
```

It seems that something was racing to create the data directory,
and so the pod would sometimes try to re-join a non-existant cluster
instead of initializing one for the first time. This tweaks the
check to hopefully be more robust. It seems to allow the cluster to
start reliably on my machine, so seems worth committing.