explicitly assert watch mod revision behavior
Etcd guarantees that, by default, watch events for a given mod revision
will never be split across multiple watch responses. This commit adds
some explicit checks for that behavior so that it's more clear.

Additionally, this changes how we track the starting revision for the
watch, to go back to starting at the _next_ expected revision rather
than the previous one.