keyspace: remove lastModRevision variable
Instead, just use `nextHeader.Revision`.

The `lastModRevision` variable can cause us to false-negative the revision
consistency check on the very first of an applied slice of WatchResponses.

Also, remove lastModRevision error because it's the same montonicity check
as looking at the first event's revision. Instead, add a sanity check
that ModRevisions are not decreasing (as guaranteed by Etcd).

Also tweak some names and typos. `lastHeader` => `nextHeader`, because
it's accumulating the next header of the keyspace.