message: Sequencer incrementally tracks pending partial sequences
If there are a large number of tracked message producers,
Sequencer.HasPending() can become a hot loop since it's very frequently
run. It's slow because it walks every journal producer to see if it has
a larger offset than the last checkpoint offset.

Instead, maintain a set of *partialSeq which begin a new transactional
sequence, and clear each *partialSeq upon dequeue of a sequence or
rollback. Or, clear _all_ pending sequences upon a Checkpoint.