gazctl shards recover: tolerate missing hints
Updates the `gazctl shards recover` subcommand to tolerate missing FSM Hints.
Previously, this command would fail if it could not find primary hints for the
given shard.  This was not necessary, since recovery is designed to work
without any hints.  This changes the `shards recover` subcommand to use the
same logic that's used in `recovery.go` for resolving a final set of hints,
which may be empty except for the journal name.  Now, you can run the command
on a shard with only backup hints, or no hints at all, and it will still
recover an identical state.