gazctl shards prune: skip journals without hints
The `shards prune` subcommand was failing with an error if any selected shard
was missing backup hints. This updates the subcommand to instead skip pruning
any journals where a shard that uses it is missing hints.

Put another way, if any selected shard is missing hints for a given recovery
log journal, we do not prune that journal.