gazctl: re-use shard and journal clients
This fixes an old bug in `shards prune` where it failed to use the
`--broker.address` argument when creating the client.

In addition, this cleans up most of the places where gazctl subcommands were
creating multiple new clients for a single operation, and consolidates down to
a single client of each type per program invocation.