gazctl: shard pruning should consider all hints
If a shard is in a persistent state where it immediately becomes FAILED
after recovery, it's very possible for the oldest recovered backup hints
to actually be _newer_ than primary hints.

Then, if we prune without considering the primary hints, the shard is
unable to recover from its log.

Fortunately this is easily remedied -- delete the primary hints while
leaving backup recovery hints in place -- but better to not do it in the
first place.