broker/client: make ErrJournalNotFound a terminal error
Currently, AppendService will loop forever waiting for an appended
journal to spring (back) into existence if it's not found.

Instead, immediately fail the AsyncAppend and all of its dependents.

This is desirable because journal deletions are often asynchronous and
race deployed shards, and there's not much that can be done aside from
failing and handling the error at a higher level.