SegmentSet.Add allow missing LastOffset
Allows adding a "covering" segment to a `SegmentSet`. Given an non-empty set,
this permits adding a segment that completely covers a subset of existing
segments and has `LastOffset: 0`.  This is needed in order to reduce backup and
primary hints into the same set, which is needed by `shards prune` in order to
prune recovery logs conservatively.