gazctl shards prune: check fragments against each hint segment separately
When pruning recovery logs,  keep a slice of all the segments from each set of hints.
Then check each segment separately, pruning only fragments that are not overlapped by
any of them. This works around difficulties we've had in the past with merging
segments from different hints into a single `SegmentSet` (which is not what
it was designed for).