gazctl shards prune: fail fast instead of pruning unpersisted fragments
If someone accidentally runs `shards prune` with a selector that doesn't
include all the shards that use a forked recovery log, it's possible for it to
try to prune the current fragment for a journal. This adds a check for that
condition so that we can provide visibility and stop the prune operation.

Attempting to prune an unpersisted fragment would fail anyway, but this
at least gives a clearer error message.