gazctl shards prune: improve audit logging
Adds a lot more information to the logs when fragments are deleted. The intent
is to run with JSON logs, to enable automated audits and analysis of prune
operations.  Because we're logging out so much more information as part of the
logs, I also turned down the log level on some of the other messages.

I also added in some additional warning logs for conditions that should never
be encountered.