gazctl shards prune: continue pruning after failure to delete fragments
This allows recovery log pruning to continue after encountering an error
removing a fragment.  We now operate Gazette clusters that use a variety of
different storage buckets, and it seems unavoidable that some of them might
have permissions misconfigured or return an error for some other reason.  In
that case, we'll now log a warning and continue the prune operation. Gazctl
will still exit non-zero if it has encountered any errors removing fragments,
to ensure it never fails silently.