journal client: force using http1.1 when fetching fragments from cloud storage
We've seen some unfortunate issues with the http2 client, where certain types
of transport failures can cause a connection to be unusable, and yet the client
will still re-use it for new http2 streams. Disabling http2 is intended to be a
short term fix, until the root cause can be identified and fixed.