Revert "journal client: force using http1.1 when fetching fragments from cloud storage"
This reverts commit 81d0aa512c02561480edea2a89f3f360d9a4a1d9.
The intent of that commit was to work around an issue where a bad connection
could be reused for subsequent requests. But I now think that this brings too
great a risk for performance degradation due to the much lower effective
connection limits. And I also suspect it may not even solve the problem, since
http1.1 connections are still re-used in many cases. So the plan is to try
something different in the next commit.