Arize dev/workload identity support (#6)
Do not error out if when we find our credentials they did not come from a json file containing a gcp service account. For workload identity FindDefaultCredentials() will not fill in the creds.JSON field but rather will attach a creds.TokenSource directly that can be used by storage.NewClient().

We leave s.signedURLOptions empty which does not seem to affect the operation of SignGet() function at top of the file.