broker: List RPC is now server-streaming and supports long-lived watches
The List RPC was originally a unary API, which has two issues:
* In large deployments, unary responses bump against gRPC maximum messages sizes.
* Unary responses can't represent a long-lived watch of changes.

Often callers want to react to changes in journals, such as reading a
newly created journal or spreading appended messages across the current
set of journals.

client.NewPolledList _somewhat_ enabled this functionality, but is
inefficient and requires selection of a polling interval.

To address both issues, List is updated to become a server-streaming API
while retaining backward comatibility with MOST clients expecting a
unary response, so long as the response size is under one thousand
journals.

The List RPC is also extended with a Watch mode. When enabled, brokers
will continue to stream real-time updates upon every change to the set
of journals which match the selector. This lets clients react instantly
to changes in journal topology.

client.NewPolledList becomes client.NewWatchedList, with a nearly
identical API (the polling interval is removed).

The shard List RPC is not modified by this change, but could follow a
similar path in the future.

Refactor pbx.NewUnroutedHeader, which is a foot-gun because it
obtains a re-entrant lock (which is not supported). Instead, make it the
caller's responsibility to obtain an RLock prior to the call.