broker: add support for fine-grain Authorizations
Authorization is implemented in terms of LabelSelectors and
Capabilities.

In order for a client to call an API, the client must present an
Authorization with a suitable Claim having:

 - The required Capability, such as READ, LIST, or APPEND.

 - A LabelSelector which scopes the resources visible to the client.

Journals which do not match the authorized Selector are not visible to
the client and are indistinguishable from not existing at all.

Authorizer and Verifier interfaces are introduced which can implement
varying strategies for obtaining and checking Authorization tokens.

AuthJournalClient drives an Authorizer to obtain and attach a suitable
credential prior to starting an RPC. AuthJournalServer uses a Verifier
to verify a caller's claims prior to dispatching a service handler.

A new `auth` module introduces NewKeyedAuth() which returns an
Authorizer and Verifier implemented in terms of pre-shared symmetric
keys. Symmetric keys are a good fit for Gazette given its distributed
nature and the requirement that brokers be able to self-sign
peer-to-peer requests, such as Replicate and proxied Read or Append
RPCs.

Client applications may want to use alternative Authorizer
implementations, such as an authorization server, but that's out of
scope for Gazette proper.