consumer: add support for fine-grain Authorizations
This change closely parallels the corersponding change to the Gazette
broker and uses much of the same infrastructure.

Built-in Shard APIs now expect and verify Authorizations which carry a
set of capabilities and a label selector which scopes the resources
(shards) which are authorized to the caller.

Shards which don't match the claim's selector are not visible to the
client and are indistinguishable from not existing at all.

AuthShardClient drives an Authorizer to obtain and attach a suitable
credential prior to starting an RPC. AuthShardServer uses a Verifier
to verify a caller's claims prior to dispatching a service handler.

Custom and application-specific APIs will likely want to use the
Service.Authorizer and Service.Verifier fields to power their own
authorizing client and server middleware.