broker/client: surface JOURNAL_NOT_FOUND from RetryReader
JOURNAL_NOT_FOUND is generally not a retry-able condition. It was
orignally made so with a "shrug", on the premise that users may create
journals after they had started their consumers. In practice,
I'm not aware of any utility for it.

Also tweak logging to reduce noise in environments that have many
journals with no current append volume, where read re-authorizations
are somewhat frequently obtained.