broker/client: quiet more noisy logging
Don't warn on a clean server-side close of a List RPC that doesn't ever
return a response. This is expected for a resumed listing which doesn't
change.

Use an attached Route of a previously completed Read, if available,
and more importantly: clear the last Response of the previous stream.
We could enter a condition where this was never cleared (for example, on
NOT_JOURNAL_BROKER) if the _following_ request went to the correct
broker, stayed open for a while with no data, and was then closed server-side.
This can cause mis-leading and false-positive logged warnings.

Update RetryReader to account for attempt being incremented on the next
loop iteration.