broker/protocol: allow `replication` of zero
Replication of zero means the journal is notational -- it's
JournalSpec is retained and it's served in listings -- but it has no
assignments and uses no broker resources.

Gazette has traditionally prohibited this, as it implies a loss of
journal consistency, but there are compelling use cases of replication:
zero when paired with read-only flags and an external mechansism to
restore journal consistency once the journal is brought back.