broker: allow an explicit append offset to exceed the max index offset
To date, we've required that this offset exactly equal the maximum index
offset, and it's been used only for `gazctl journal reset-head`.

Estuary has another interesting use case: we intend to use zero-byte
appends to restore the write head of a "suspended" journal, which is a
read-only journal that may have had a replication factor of zero (inplying
lost journal consistency), and which many not have any fragments in the
index (due to bucket lifecycle policy), but for which a suspended
write-head was retained as an application-specific label attached
to the JournalSpec.