broker: implement journal suspension
Introduce a journal suspension mechanism which is incorporated into the
Append RPC state machine, and stored as a `suspend` field within each
JournalSpec.

Suspended journals consumer fewer or no assignment slots, reducing
broker and Etcd resources required for journals which are not actively
serving appends.

When running with the `--auto-suspend` flag, brokers will automatically
suspend journals which are fully flushed to the fragment index.